+++ src/LightOps.Commerce.Services.Navigation/Api/Queries/FetchNavigationsByHandlesQuery.cs
@@ -3,7 +3,7 @@ using LightOps.CQRS.Api.Queries;
 
 namespace LightOps.Commerce.Services.Navigation.Api.Queries
 {
-    public class FetchNavigationsByHandleQuery : IQuery
+    public class FetchNavigationsByHandlesQuery : IQuery
     {
         public IList<string> Handles { get; set; }
     }