+++ src/LightOps.Commerce.Services.Navigation/Api/Queries/FetchNavigationsByIdsQuery.cs
@@ -3,7 +3,7 @@ using LightOps.CQRS.Api.Queries;
 
 namespace LightOps.Commerce.Services.Navigation.Api.Queries
 {
-    public class FetchNavigationsByIdQuery : IQuery
+    public class FetchNavigationsByIdsQuery : IQuery
     {
         public IList<string> Ids { get; set; }
     }