+++ src/LightOps.Commerce.Services.Navigation/Api/QueryHandlers/IFetchNavigationsByIdsQueryHandler.cs
@@ -5,7 +5,7 @@ using LightOps.CQRS.Api.Queries;
 
 namespace LightOps.Commerce.Services.Navigation.Api.QueryHandlers
 {
-    public interface IFetchNavigationsByIdQueryHandler : IQueryHandler<FetchNavigationsByIdQuery, IList<INavigation>>
+    public interface IFetchNavigationsByIdsQueryHandler : IQueryHandler<FetchNavigationsByIdsQuery, IList<INavigation>>
     {
 
     }