+++ src/LightOps.Commerce.Services.Navigation/Configuration/INavigationServiceComponent.cs
@@ -21,8 +21,8 @@ namespace LightOps.Commerce.Services.Navigation.Configuration
         INavigationServiceComponent OverrideCheckNavigationHealthQueryHandler<T>() where T : ICheckNavigationHealthQueryHandler;
         INavigationServiceComponent OverrideFetchNavigationsByParentIdQueryHandler<T>() where T : IFetchNavigationsByParentIdQueryHandler;
         INavigationServiceComponent OverrideFetchNavigationsByRootQueryHandler<T>() where T : IFetchNavigationsByRootQueryHandler;
-        INavigationServiceComponent OverrideFetchNavigationsByHandleQueryHandler<T>() where T : IFetchNavigationsByHandleQueryHandler;
-        INavigationServiceComponent OverrideFetchNavigationsByIdQueryHandler<T>() where T : IFetchNavigationsByIdQueryHandler;
+        INavigationServiceComponent OverrideFetchNavigationsByHandlesQueryHandler<T>() where T : IFetchNavigationsByHandlesQueryHandler;
+        INavigationServiceComponent OverrideFetchNavigationsByIdsQueryHandler<T>() where T : IFetchNavigationsByIdsQueryHandler;
         INavigationServiceComponent OverrideFetchNavigationByHandleQueryHandler<T>() where T : IFetchNavigationByHandleQueryHandler;
         INavigationServiceComponent OverrideFetchNavigationByIdQueryHandler<T>() where T : IFetchNavigationByIdQueryHandler;
         #endregion Query Handlers